Rabbit terminado a falta de probar, se empieza Consumidores el CF4 está a falta de mostrar la información y probar

+++ ApiInformers/src/main/java/com/api_informers/controller/FileController.java
@@ -47,7 +47,7 @@ public class FileController {
 	@PostMapping(value="/newFile")
 	public void createPost(@RequestParam MultipartFile file, @RequestParam String title, 
 										   @RequestParam String description, @RequestParam List<String> keywords,
-										   @RequestParam Integer size,
+										   @RequestParam Double size,
 										   HttpServletRequest request) throws IOException {
 		
 		//Comprobar que el archivo subido es de formato JSON
